Add files to chat rather than auto-summarising
Files are now added to the chat itself. Previously they were sent to
the model in a separate chat, and the content of the file didn't end
up in the chat.

This new approach allows users to choose the questions to ask of
the uploaded file, and, because the file content is in the chat, they
can ask various questions about the file.

Overall this simplifies the code a lot, and also improves the UX by
allowing more flexibility for the user. Win!

More files can be added to the chat, if they are small.