Fix duplicate UI displaying on first message of new chat
Not quite sure why this was happening, but removing the "Success" banner
stopped the existing UI not disappearing correctly when writing out the
first new message after clicking new chat.

I guessed it might be the success banner because the problem didn't
happen on the very first chat after refreshing the page.